gbw/requirements.txt
PySimpleGUI
subprocess
pillow
qrcode
os
tkinter
time
threading
playsound
flask